rename nearbypoistreams: new became the normal one, original became original test
