+++ common/src/main/java/net/caffeinemc/mods/lithium/common/world/interests/iterator/NearbyPointOfInterestStream.java
@@ -1,7 +1,8 @@
 
 package net.caffeinemc.mods.lithium.common.world.interests.iterator;
 
-import it.unimi.dsi.fastutil.longs.LongArrayList;
+import it.unimi.dsi.fastutil.longs.LongIterator;
+import it.unimi.dsi.fastutil.objects.ObjectArrayList;
 import net.caffeinemc.mods.lithium.common.util.Distances;
 import net.caffeinemc.mods.lithium.common.util.tuples.SortedPointOfInterest;
 import net.caffeinemc.mods.lithium.common.world.interests.PointOfInterestSetExtended;
@@ -16,10 +17,7 @@ import net.minecraft.world.entity.ai.village.poi.PoiType;
 import net.minecraft.world.level.ChunkPos;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.Spliterator;
-import java.util.Spliterators;
+import java.util.*;
 import java.util.function.Consumer;
 import java.util.function.Predicate;
 
@@ -35,14 +33,30 @@ public class NearbyPointOfInterestStream extends Spliterators.AbstractSpliterato
     private final Predicate<Holder<PoiType>> typeSelector;
     private final PoiManager.Occupancy occupationStatus;
 
-    private final LongArrayList chunksSortedByMinDistance;
     private final ArrayList<SortedPointOfInterest> points;
     private final Predicate<PoiRecord> afterSortingPredicate;
     private final Consumer<PoiRecord> collector;
     private final BlockPos origin;
 
-    private int chunkIndex;
-    private double currChunkMinDistanceSq;
+    private final int clampedOriginChunkY;
+
+    private final int chunkYMin;
+    private final double minChunkYDistSq;
+
+    private final ObjectArrayList<QueuedSection> queuedPOISections;
+    private int queuedSectionsSearched;
+    private boolean forciblyDeplete;
+    private final int forciblyDepleteTrigger;
+    private int ring;
+    private final int ringMax;
+    private final LongIterator ringIterator;
+
+    private final int ringClosestEdgeDistance;
+    private double closestRingDistanceSq;
+
+    private double nextSectionDistanceSq;
+    private double lowestWaitingDistanceSq;
+    private final double distanceLimitL2Sq;
     private int pointIndex;
     private final Comparator<? super SortedPointOfInterest> pointComparator;
 
@@ -57,14 +71,56 @@ public class NearbyPointOfInterestStream extends Spliterators.AbstractSpliterato
 
         this.storage = storage;
 
-        this.chunkIndex = 0;
-        this.pointIndex = 0;
-
         this.points = new ArrayList<>();
         this.occupationStatus = status;
         this.typeSelector = typeSelector;
 
         this.origin = origin;
+        this.chunkYMin = this.storage.lithium$getChunkYMin();
+        final int chunkYMax = this.storage.lithium$getChunkYMaxInclusive();
+
+        this.clampedOriginChunkY = Math.clamp(SectionPos.blockToSectionCoord(origin.getY()), chunkYMin, chunkYMax);
+
+        // If the origin is outside of build limit (e.g. 1.18+ overworld to nether portal teleport) in the target dimension,
+        // the min Y distance is not zero for a chunk.
+        final int minChunkYDist = Math.min(Math.max(this.origin.getY(), SectionPos.sectionToBlockCoord(this.chunkYMin)),
+                SectionPos.sectionToBlockCoord(chunkYMax, 15)) - this.origin.getY();
+        this.minChunkYDistSq = minChunkYDist * minChunkYDist;
+
+        final int originX = this.origin.getX();
+        final int originZ = this.origin.getZ();
+        final int chunkMaxX = SectionPos.blockToSectionCoord(origin.getX()+radius);
+        final int chunkMinX = SectionPos.blockToSectionCoord(origin.getX()-radius);
+        final int chunkMaxZ = SectionPos.blockToSectionCoord(origin.getZ()+radius);
+        final int chunkMinZ = SectionPos.blockToSectionCoord(origin.getZ()-radius);
+
+        // Chunk searches will expand in a concentric rings around the origin chunk
+        this.ring = 0;
+        final int originChunkX = SectionPos.blockToSectionCoord(originX);
+        final int originChunkZ = SectionPos.blockToSectionCoord(originZ);
+        this.ringMax = Math.max(Math.max(chunkMaxX - originChunkX, originChunkX - chunkMinX),
+                Math.max(chunkMaxZ - originChunkZ, originChunkZ - chunkMinZ));
+        this.ringIterator = getRingsOfChunksIterator(originChunkX, chunkMaxX, chunkMinX, originChunkZ, chunkMaxZ, chunkMinZ);
+
+        // Initialize Ring Distances
+        // This is used to decide the minimum distance to possible POI located in the next ring
+        this.ringClosestEdgeDistance = Math.min(Math.min((originX & 15) + 1, 16 - originX & 15),
+                Math.min((originZ & 15) + 1, 16 - originZ & 15));
+        this.closestRingDistanceSq = this.getPotentialRingDistanceSq();
+
+        this.nextSectionDistanceSq = Double.MAX_VALUE;
+        this.lowestWaitingDistanceSq = Double.MAX_VALUE;
+
+        // Note: This is much faster than PriorityHeapQueue because dequeues become very expensive
+        // Also keep track of square distances because otherwise comparisons become very expensive
+        // Todo: If and when value records are thing convert over to it
+        final int sectionsPerChunk = chunkYMax - chunkYMin + 1;
+        final int listSize = sectionsPerChunk * 9; // Sufficient to store 2 full rings [ring: 0-1]
+        this.queuedPOISections = new ObjectArrayList<>(listSize);
+        this.queuedSectionsSearched = 0; // Also the index for the next queued section
+        this.forciblyDeplete = false;
+        this.forciblyDepleteTrigger = listSize - sectionsPerChunk; // Leave 1 chunk's worth left to avoid resizing
+
         if (useSquareDistanceLimit) {
             this.collector = (point) -> {
                 if (Distances.isWithinSquareRadius(this.origin, radius, point.getPos())) {
@@ -80,8 +136,7 @@ public class NearbyPointOfInterestStream extends Spliterators.AbstractSpliterato
             };
         }
 
-        double distanceLimitL2Sq = useSquareDistanceLimit ? radius * radius * 2 : radius * radius;
-        this.chunksSortedByMinDistance = initChunkPositions(origin, radius, distanceLimitL2Sq);
+        this.distanceLimitL2Sq = useSquareDistanceLimit ? radius * radius * 2 : radius * radius;
         this.afterSortingPredicate = afterSortingPredicate;
         this.pointComparator = preferNegativeY ? (o1, o2) -> {
             // Use the cached values from earlier
@@ -126,35 +181,6 @@ public class NearbyPointOfInterestStream extends Spliterators.AbstractSpliterato
         };
     }
 
-    private static LongArrayList initChunkPositions(BlockPos origin, int radius, double distanceLimitL2Sq) {
-        int minChunkX = (origin.getX() - radius - 1) >> 4;
-        int minChunkZ = (origin.getZ() - radius - 1) >> 4;
-
-        int maxChunkX = (origin.getX() + radius + 1) >> 4;
-        int maxChunkZ = (origin.getZ() + radius + 1) >> 4;
-
-        LongArrayList chunkPositions = new LongArrayList();
-
-        // todo: Find a better way to go about this that doesn't require allocating a ton of positions
-        for (int chunkX = minChunkX; chunkX <= maxChunkX; chunkX++) {
-            for (int chunkZ = minChunkZ; chunkZ <= maxChunkZ; chunkZ++) {
-                if (distanceLimitL2Sq >= Distances.getMinChunkToBlockDistanceL2Sq(origin, chunkX, chunkZ)) {
-                    chunkPositions.add(ChunkPos.asLong(chunkX, chunkZ));
-                }
-            }
-        }
-
-        // Sort all the chunks by their distance to the search origin. The points in each chunk are sorted independently
-        // in their own bucket (the chunk itself), but this does not change the ordering of points between each bucket.
-        chunkPositions.sort(
-                (long c1, long c2) -> Double.compare(
-                        Distances.getMinChunkToBlockDistanceL2Sq(origin, ChunkPos.getX(c1), ChunkPos.getZ(c1)),
-                        Distances.getMinChunkToBlockDistanceL2Sq(origin, ChunkPos.getX(c2), ChunkPos.getZ(c2))));
-
-        return chunkPositions;
-    }
-
-
     @Override
     public boolean tryAdvance(Consumer<? super PoiRecord> action) {
         // Accepted POI:
@@ -165,39 +191,35 @@ public class NearbyPointOfInterestStream extends Spliterators.AbstractSpliterato
         // return the one with most negative X. If several exist:
         // return the one with most negative Z. If several exist: Be confused about two POIs being in the same location.
 
-        // Check to see if we still have points to return
         if (this.pointIndex < this.points.size()) {
             if (this.tryAdvancePoint(action)) {
                 return true;
             }
         }
 
-        // Find the next ordered chunk to scan for points
-        while (this.chunkIndex < this.chunksSortedByMinDistance.size()) {
-            long chunkPos = this.chunksSortedByMinDistance.getLong(this.chunkIndex);
-            int chunkPosX = ChunkPos.getX(chunkPos);
-            int chunkPosZ = ChunkPos.getZ(chunkPos);
+        while (this.ringIterator.hasNext() || !this.isSectionListEmpty()) {
+            this.keepAddingRingsUntilSufficient();
 
-            // Keep track of the guaranteed minimum distance of newly collected POI. We can only consume the closest
-            // POI if we know that there will not be any other POI with a smaller distance.
-            this.currChunkMinDistanceSq = Distances.getMinChunkToBlockDistanceL2Sq(this.origin, chunkPosX, chunkPosZ);
-            this.chunkIndex++;
-            if (this.chunkIndex == this.chunksSortedByMinDistance.size()) {
-                this.currChunkMinDistanceSq = Double.POSITIVE_INFINITY;
+            final int previousSize = this.points.size();
+            while (!this.isSectionListEmpty() && this.lowestWaitingDistanceSq >= this.getMinimumNextPotentialDistanceSq()) {
+                final long sectionPos = queuedPOISections.get(this.queuedSectionsSearched++).sectionPos;
+                this.nextSectionDistanceSq = this.getNextSectionDistanceSq();
+                final Optional<PoiSection> poiSection = this.storage.lithium$getElementAt(sectionPos);
+                if (poiSection.isPresent()) {
+                    ((PointOfInterestSetExtended)poiSection.get())
+                            .lithium$collectMatchingPoints(this.typeSelector, this.occupationStatus, this.collector);
+                }
+
+                this.forciblyDeplete = (!this.forciblyDeplete || !this.isSectionListEmpty()) && this.forciblyDeplete;
+                if(this.points.size() > previousSize) {
+                    this.points.subList(this.pointIndex, this.points.size()).sort(this.pointComparator);
+                    break;
+                }
+
+                if(!forciblyDeplete && this.nextSectionDistanceSq > this.closestRingDistanceSq) {
+                    break;
+                }
             }
-
-            int previousSize = this.points.size();
-            // Collect all points in the chunk into the active list of points
-            for (PoiSection set : this.storage.lithium$getInChunkColumn(chunkPosX, chunkPosZ)) {
-                ((PointOfInterestSetExtended) set).lithium$collectMatchingPoints(this.typeSelector, this.occupationStatus, this.collector);
-            }
-
-            // If no points were found in this chunk, skip it early and move on
-            if (this.points.size() == previousSize) {
-                continue;
-            }
-
-            this.points.subList(this.pointIndex, this.points.size()).sort(this.pointComparator);
 
             // Return the first point in the chunk
             if (this.tryAdvancePoint(action)) {
@@ -205,19 +227,17 @@ public class NearbyPointOfInterestStream extends Spliterators.AbstractSpliterato
             }
         }
 
-        // Return the first valid point
         return this.tryAdvancePoint(action);
-        //Returns true when progress was made by consuming an element
-        //Returns false when no progress was made because no more elements exist.
     }
-
 
     private boolean tryAdvancePoint(Consumer<? super PoiRecord> action) {
         while (this.pointIndex < this.points.size()) {
             SortedPointOfInterest next = this.points.get(this.pointIndex);
 
-            //Only consume points if we are sure that there are no closer (or same distance) points to be scanned. Otherwise scan more chunks
-            if (next.distanceSq() >= this.currChunkMinDistanceSq) {
+            // Only consume points if we are sure that there are no closer (or same distance) points to be scanned.
+            // Otherwise, scan more chunks
+            if (next.distanceSq() >= this.getMinimumNextPotentialDistanceSq()) {
+                this.lowestWaitingDistanceSq = next.distanceSq();
                 return false;
             }
             this.pointIndex++;
@@ -231,4 +251,140 @@ public class NearbyPointOfInterestStream extends Spliterators.AbstractSpliterato
         return false; //No more points. Scan more chunks
     }
 
+    /*
+     * Keep searching for POI section in concentric rings of chunks around the origin. Continue if:
+     * List is empty
+     * There are more rings to scan.
+     * The closer of:
+     * - The next section
+     * - The closest POI in the point list
+     * is not closer than:
+     * - Unchecked possible sections in the list
+     * - Chunks in next ring
+     */
+    private void keepAddingRingsUntilSufficient() {
+        if (!this.forciblyDeplete && this.ringIterator.hasNext() &&
+                (Math.min(this.lowestWaitingDistanceSq, this.nextSectionDistanceSq) >= this.closestRingDistanceSq)) {
+            this.queuedPOISections.removeElements(0, this.queuedSectionsSearched);
+            this.queuedSectionsSearched = 0;
+            int ringStart = this.ring;
+            do {
+                final long chunkPos = this.ringIterator.nextLong();
+                final int currentChunkX = ChunkPos.getX(chunkPos);
+                final int currentChunkZ = ChunkPos.getZ(chunkPos);
+
+                if (this.distanceLimitL2Sq >= Distances.getMinChunkToBlockDistanceL2Sq(this.origin, currentChunkX, currentChunkZ)) {
+                    // This iterates the column in order of the closest section
+                    // Doing so noticeably reduces sorting duration later on because parts of the list will already be sorted
+                    final BitSet poiSections = this.storage.lithium$getNonEmptyPOISections(currentChunkX, currentChunkZ);
+                    int upperBit = poiSections.nextSetBit(this.clampedOriginChunkY - this.chunkYMin);
+                    int upperDist = this.getYDistanceFromBitIndex(upperBit);
+                    int lowerBit = poiSections.previousSetBit(this.clampedOriginChunkY - this.chunkYMin - 1);
+                    int lowerDist = this.getYDistanceFromBitIndex(lowerBit);
+                    while (upperBit != -1 || lowerBit != -1) {
+                        final int currentChunkY;
+                        if(lowerDist <= upperDist) {
+                            currentChunkY = lowerBit + this.chunkYMin;
+                            lowerBit = lowerBit == -1 ? -1 : poiSections.previousSetBit(lowerBit-1);
+                            lowerDist = this.getYDistanceFromBitIndex(lowerBit);
+                        } else {
+                            currentChunkY = upperBit + this.chunkYMin;
+                            upperBit = upperBit == -1 ? -1 : poiSections.nextSetBit(upperBit+1);
+                            upperDist = this.getYDistanceFromBitIndex(upperBit);
+                        }
+                        this.queuedPOISections.add(
+                                new QueuedSection(
+                                        SectionPos.asLong(currentChunkX, currentChunkY, currentChunkZ),
+                                        Distances.getMinSectionDistanceSq(
+                                                this.origin, currentChunkX, currentChunkY, currentChunkZ)
+                                )
+                        );
+                    }
+
+                    this.forciblyDeplete = this.queuedPOISections.size() > this.forciblyDepleteTrigger;
+                }
+
+                if (forciblyDeplete || this.ring > ringStart) {
+                    this.sortSectionList();
+                    this.nextSectionDistanceSq = this.getNextSectionDistanceSq();
+                    if(forciblyDeplete
+                            || Math.min(this.lowestWaitingDistanceSq, this.nextSectionDistanceSq)
+                            < this.closestRingDistanceSq) {
+                        break;
+                    }
+                    ringStart = this.ring;
+                }
+            } while (this.ringIterator.hasNext());
+        }
+    }
+
+    private void sortSectionList() {
+        // Note: Do not use unstable sort - the fastutils quicksort is quite a bit slower
+        queuedPOISections.subList(this.queuedSectionsSearched, this.queuedPOISections.size())
+                .sort(Comparator.comparingDouble(qS -> qS.minDistance));
+    }
+
+    private boolean isSectionListEmpty(){
+        return this.queuedPOISections.size() <= this.queuedSectionsSearched;
+    }
+
+    // Minimum of the next [closest] section in the queue or the closest potential unchecked chunks [next ring]
+    private double getMinimumNextPotentialDistanceSq() {
+        return Math.min(this.nextSectionDistanceSq, this.closestRingDistanceSq);
+    }
+
+    private double getNextSectionDistanceSq() {
+        return this.isSectionListEmpty() ?
+                Double.MAX_VALUE : this.queuedPOISections.get(this.queuedSectionsSearched).minDistance;
+    }
+
+    private int getYDistanceFromBitIndex(final int bitIndex) {
+        return bitIndex == -1 ?
+                Integer.MAX_VALUE :
+                Math.abs(Distances.getClosestAlongSectionAxis(
+                        this.origin.getY(), bitIndex + this.chunkYMin) - this.origin.getY());
+    }
+
+    // Expand chunks to search in concentric square rings
+    // Todo: Gap-less circle drawing algorithms may have better performance
+    private double getPotentialRingDistanceSq() {
+        final int ringDistance = Math.max(this.ring - 1,0) * 16 + (ring > 0 ? this.ringClosestEdgeDistance : 0);
+        return this.ring > this.ringMax ? Double.MAX_VALUE : ringDistance * ringDistance + this.minChunkYDistSq;
+    }
+
+    private LongIterator getRingsOfChunksIterator(final int cx, final int maxX, final int minX,
+                                                  final int cz, final int maxZ, final int minZ) {
+        return new LongIterator() {
+            int x = 0;
+            int fx = cx;
+            int z = 0;
+            int fz = cz;
+            @Override
+            public long nextLong() {
+                long res = ChunkPos.asLong(fx, fz);
+                do {
+                    z = z > 0 ? -z : 1 - z;
+                    if (z > ring) {
+                        x = x > 0 ? -x : 1 - x;
+                        if (x > ring) {
+                            x = 0;
+                            ring++;
+                            closestRingDistanceSq = getPotentialRingDistanceSq();
+                        }
+                        z = x < ring && x > -ring ? ring : 0;
+                    }
+                    fx = cx + x;
+                    fz = cz + z;
+                } while (ring <= ringMax && fx < minX || fx > maxX || fz < minZ || fz > maxZ);
+                return res;
+            }
+
+            @Override
+            public boolean hasNext() {
+                return ring <= ringMax;
+            }
+        };
+    }
+
+    private record QueuedSection(long sectionPos, double minDistance) {}
 }
